+++ src/setupClaCheck.ts
@@ -22,7 +22,7 @@ export async function setupClaCheck() {
   let committerMap = getInitialCommittersMap()
 
   let committers = await getCommitters()
-  committers = checkAllowList(committers)
+  committers = await checkAllowList(committers)
   committers = await applyOrgExemption(committers)
 
   const { claFileContent, sha } = (await getCLAFileContentandSHA(